Move trace package to utils.
Also moved edu.snu.cay.utils directory to edu/snu/cay/utils.
