+++ services/elastic-memory/src/main/avro/elasticmemory.avsc
@@ -125,7 +125,7 @@
     {"name": "srcId", "type" : "string"},
     {"name": "destId", "type" : "string"},
     {"name": "operationId", "type" : ["null", "string"], "default": null},
-    {"name": "traceInfo", "type": ["null", "edu.snu.cay.services.em.trace.AvroTraceInfo"], "default": null},
+    {"name": "traceInfo", "type": ["null", "edu.snu.cay.utils.trace.AvroTraceInfo"], "default": null},
     {"name": "ctrlMsg", "type": ["null", "CtrlMsg"], "default": null},
     {"name": "dataMsg", "type": ["null", "DataMsg"], "default": null},
     {"name": "resultMsg", "type": ["null", "ResultMsg"], "default": null},
