+++ services/elastic-memory/src/main/java/edu/snu/cay/services/em/driver/ElasticMemoryMsgHandler.java
@@ -18,7 +18,7 @@ package edu.snu.cay.services.em.driver;
 import edu.snu.cay.services.em.avro.AvroElasticMemoryMessage;
 import edu.snu.cay.services.em.avro.RegisMsg;
 import edu.snu.cay.services.em.msg.api.ElasticMemoryCallbackRouter;
-import edu.snu.cay.services.em.trace.HTraceUtils;
+import edu.snu.cay.utils.trace.HTraceUtils;
 import edu.snu.cay.utils.SingleMessageExtractor;
 import org.htrace.Trace;
 import org.htrace.TraceScope;
