+++ utils/src/main/java/edu/snu/cay/utils/AvroUtils.java
