+++ utils/src/main/java/edu/snu/cay/utils/LongRangeUtils.java
