+++ utils/src/main/java/edu/snu/cay/utils/SingleMessageExtractor.java
