+++ utils/src/main/java/edu/snu/cay/utils/StateMachine.java
