+++ utils/src/main/java/edu/snu/cay/utils/trace/HTrace.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.cay.services.em.trace;
+package edu.snu.cay.utils.trace;
 
 import org.htrace.Sampler;
 import org.htrace.SpanReceiver;
